==> notebooks/01-Evaluation.ipynb <==
# Evaluation

## How To Evaluate
You need to run two nodejs scripts in sequence to collect accessibility evaluation results:

1. `../javascript/evaluate-and-save-reports.js`
2. `../javascript/summarize-reports.js`

Before running, several things prepare:

1. Under the `../data` folder, create a new folder with a meaningful date (e.g., `07-29-2024`)
2. Add csv file(s) under the `../data/input` that lists URLs that you want to test. The script will load all CSV files (`*.csv`) and run the evaluation.
3. The csv file at minimum should contain a column name `url` (non-headless)
4. Other columns used, if present, are `page_type` (e.g., `home` or `document`) and `page_id` (an unique ID to this page)

## Run Evaluation
### Configure the Parameter

"""
The name of the folder you created under `../data/`
"""
TIME_STAMP_FOLDER_NAME = '2024-08-01'

"""
Let's check if an input file exist
"""
!ls -R ../data/{TIME_STAMP_FOLDER_NAME}/input

### Run the Testing

"""
Run the accessibility testing. This takes 10 hours for 7K+ URLs.
"""
!node ../javascript/evaluate-and-save-reports.js {TIME_STAMP_FOLDER_NAME}

### Summarize the reports

"""
Summarize the reports and save a CSV file. This shouldn't take more than a minute.
"""
!node --max-old-space-size=8192 ../javascript/summarize-reports.js {TIME_STAMP_FOLDER_NAME}

"""
Check if files are created properly
"""
!ls -R ../data/{TIME_STAMP_FOLDER_NAME}/reports
