==> notebooks/04-Manual-Testing-Results.ipynb <==
import altair as alt
import pandas as pd

# Define custom theme to be applied to all plots
def theme():
    return {
        "config": {
            "title": {
                "dy": 1,
                "fontSize": 22,
                "fontWeight": 400,
                "align": "center",
                "anchor": "middle",
                "subtitleColor": "grey",
                "subtitleFontSize": 22
            },
            "view": {
                "fill": "#FCFCFC",
            },
            "header": {
                "titleFontSize": 23,
                "labelFontSize": 23,
                "labelFontWeight": 400,
            },
            "axis": {
                "titleFontSize": 23,
                "titleFontWeight": 400,
                "labelFontSize": 20,
                "labelFontWeight": 400,
                "labelLimit": 1000,
                "domainWidth": 1.5,
                "domainColor": "black",
                "tickColor": "black",
                "domain": False
            },
            "axisX": {
                "domain": True
            },
            "legend": {
                "titleFontSize": 23,
                "titleFontWeight": 400,
                "labelFontSize": 23,
                "labelLimit": 1000,
                "strokeColor": '#F4F6F7',
                "padding": 15
            }
        }
    }

alt.themes.register("theme", theme)
alt.themes.enable("theme")

"""
The name of the folder you created under `../data/`
"""
TIME_STAMP_FOLDER_NAME = '2024-08-01'

def save_figure(plot=None, name=None, data=None):
    plot.save(f'../data/{TIME_STAMP_FOLDER_NAME}/figures/{name}.png', scale_factor=8)
    plot.save(f'../data/{TIME_STAMP_FOLDER_NAME}/figures/{name}.svg')
    plot.save(f'../data/{TIME_STAMP_FOLDER_NAME}/website/{name}.json')
    plot.display()
    if data is not None:
        data.to_csv(f'../data/{TIME_STAMP_FOLDER_NAME}/website/{name}.csv')

df = pd.read_csv(f'../data/{TIME_STAMP_FOLDER_NAME}/manual/Manual Test Results.csv')

df['TASK'] = 'T' + df.TASK_ID.astype(str) + '. ' + df.TASK_SHORT_DESC

df.SECONDS = df.SECONDS.apply(lambda x: 300 if x > 300 else x)

df['MINUTES'] = df.SECONDS
df.MINUTES /= 60

df['IS_SUCCESS'] = df.SUCCESS.apply(lambda x: 'Success' if x == 'O' else 'Fail')
df.SUCCESS = df.SUCCESS.apply(lambda x: 'Succeeded in 5 mins' if x == 'O' else 'Failed in 5 mins')

df.BASELINE = 4

df.FRUSTRATION = 8 - df.FRUSTRATION

df

TASKS_SORTED = df.groupby('TASK').agg({'SECONDS': 'mean'}).reset_index().sort_values(by='SECONDS', ascending=False).TASK.tolist()
TASKS_SORTED

df.head(3)

"""
Averate Time Taken
"""
df.SECONDS.mean() / 60

"""
Succeded Tasks
"""
df[df.IS_SUCCESS == 'Success'].IS_SUCCESS.count()

[df.CONFIDENCE.mean(), df.SATISFACTION.mean(), df.FRUSTRATION.mean()]

def bar(c=None):
    df_copy = (df.copy() if c is None else df[df.DATA_PORTAL == c].copy())
    
    bar = alt.Chart(
        df_copy
    ).mark_bar(
        cornerRadiusBottomRight=3,
        cornerRadiusTopRight=3
    ).encode(
        alt.X('mean(MINUTES):Q', title=None).scale(domain=[0, 5], clamp=True).axis(format='.0f', tickCount=5),
        alt.Y('TASK:N', title='Tasks', sort=TASKS_SORTED).axis(titleX=-360, zindex=1),
        color=(alt.Color() if c is None else alt.Color('SUCCESS:N', title='Task completion').scale(range=['#D55D00', '#56B4E9']).legend(orient='bottom')),
        # alt.Row('DATA_PORTAL:N', title=None)
    ).properties(
        title={
            'text': 'Time Taken (minutes)',
            # 'subtitle': 'Data as of Jul 24, 2024',
            "dy": -10
        },
        width=400,
        height=300
    )
    return bar

bar()

def sub(p=None):
    COLS = ['CONFIDENCE', 'SATISFACTION', 'FRUSTRATION']

    _ = df.copy() if p is None else df[df.DATA_PORTAL == p].copy()
    df['baseline'] = 4
    
    plot_likert = None
    for c in COLS:
        _plot = alt.Chart(
            _
        ).mark_point(size=150, opacity=1, filled=True, color='black').encode(
            alt.X(f'mean({c}):Q', title=None).scale(domain=[1, 7] if c != 'FRUSTRATION' else [7, 1]).axis(tickCount=5), # , '(1=negative, 7=positive)'
            alt.Y('TASK:N', title='Tasks', sort=TASKS_SORTED).axis(None),
            # alt.Color('CONFIDENCE:N', title=None, legend=None).scale(domain=[1, 2, 3, 4, 5, 6, 7], range=['#56B4E9', '#56B4E9', '#56B4E9', 'grey', '#D55D00', '#D55D00', '#D55D00']),
        ).properties(
            title={
                'text': f'{c.title()} Level',
                # 'subtitle': 'Data as of Jul 24, 2024',
                "dy": -10
            },
            width=300,
            height=300
        )
    
        err = _plot.mark_errorbar().encode(
            x=alt.X(f'{c}:Q', title=[f'{c.title()}']).scale(zero=False),
            color=alt.value('grey')
        )
        
        rule = alt.Chart(
            _
        ).mark_rule(color='black', strokeDash=[3, 3]).encode(
            alt.X('baseline:Q').scale(domain=[1, 7] if c != 'FRUSTRATION' else [7, 1])
        )

        if p is None:
            _plot = rule + err + _plot
        else:
            _plot = rule + _plot
    
        # _plot = _plot.facet(row=alt.Row('DATA_PORTAL:N', title=None, header=None))
    
        if plot_likert == None:
            plot_likert = _plot
        else:
            plot_likert |= _plot
    
    plot_likert = plot_likert.resolve_scale(y='shared')
    return plot_likert

sub()

# suc = alt.Chart(df).mark_point(filled=True, size=400).encode(
#     # alt.X('SUCCESS:Q', title=''),
#     alt.Y('TASK:N', title=None, sort=TASKS_SORTED).axis(None),
#     alt.Opacity('IS_SUCCESS:N', legend=None).scale(domain=['Success', 'Fail'], range=[1, 0]),
#     # alt.Color('SUCCESS:N', title='Task completion').scale(range=['#D55D00', '#56B4E9']).legend(orient='bottom'),
#     alt.Column('DATA_PORTAL:N', title=None, spacing=1).header(labelAngle=-90, labelAnchor='end', labelAlign='right')
# ).properties(
#     title={
#         'text': 'Task Success',
#         # 'subtitle': 'Data as of Jul 24, 2024',
#         "dy": -10
#     },
#     width=40,
#     height=300
# )
# suc

# piedf = df.groupby(['TASK', 'SUCCESS']).agg({'IS_SUCCESS': 'count'}).reset_index()
# piedf.IS_SUCCESS /= 3

# pie = alt.Chart(piedf).mark_arc().encode(
#     alt.Theta('IS_SUCCESS', title='').stack(True),
#     alt.Row('TASK:N', title='Tasks', sort=TASKS_SORTED,spacing=2).header(None),#.axis(titleX=-360, zindex=1),
#     # alt.Opacity('IS_SUCCESS:N', legend=None).scale(domain=['Success', 'Fail'], range=[1, 0]),
#     alt.Color('SUCCESS:N', title='Task completion').scale(domain=['Succeeded in 5 mins', 'Failed in 5 mins'], range=['#4C78A8', 'lightgrey']).legend(None),
#     # alt.Column('DATA_PORTAL:N', title=None, spacing=1).header(labelAngle=-90, labelAnchor='end', labelAlign='right')
# ).properties(
#     title={
#         'text': 'Succeded',
#         # 'subtitle': 'Data as of Jul 24, 2024',
#         "dy": -10
#     },
#     width=40,
#     height=29
# )
# # pie#.resolve_scale(x='independent')
# # piedf
# pie

def row(portal=None):
    return alt.hconcat(
        bar(portal),
        sub(portal)
    ).resolve_scale(y='shared').properties(title=portal if portal is not None else '')

plot = alt.vconcat(
    row(),
    row('cBioPortal'),
    row('ENCODE'),
    row('HuBMAP'),
)

"""
Save for the manuscript figures and website plots
"""
save_figure(plot, 'manual-test-results', df)
